다시 실행

+++ FreeBoard/deploy-before.sh
@@ -1,3 +1,4 @@
-sudo rm -rf /home/ec2-user/build/
-
-sudo mkdir  /home/ec2-user/build/
+if [ -d /home/ec2-user/build ]; then
+    sudo rm -rf /home/ec2-user/build/
+fi
+sudo mkdir -vp /home/ec2-user/build/
